requirements.txt
fastapi
uvicorn
google-generativeai
pymongo
redis
pinecone-client
dnspython  
grpcio-tools
grpcio
python-dotenv
pytz
